// src/app/enums/catalog.enum.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum CatalogPrefix {
    DISPOSITIVOS_CONECTADOS = 'connectedDevices-',
    REBOOT_AUTOMATICO = 'rebotAutomatico-',
    REBOOT_MANUAL = 'rebootManual-',
    TROCA_CANAL = 'setChannel-',
    TROCA_APARELHOS_NAO_FAZ = 'trocaAparelhoNaoFaz-',
    VOLUME_RUIDO = 'volumeRuido-',
    CABOS_E_PORTAS = 'cabosPortas-',
    TROCA_APARELHOS_NAO_RECEBE = 'trocaAparelhosNaoRecebe-',
    TROCA_APARELHOS = 'trocaAparelhos-',
    TROCA_APARELHOS_NAO_FAZ_NAO_RECEBE = 'trocaAparelhosNaoFazNaoRecebe-',
    ISOLAR_MODEM = 'isolarModem-',
    TV_MANUAL_STB = 'tvManualStb-',
    REBOOT_AUTOMATICO_STB = 'rebootAutomaticoStb-',
    REBOOT_MANUAL_STB = 'rebootManualStb-',
    REBOOT_MANUAL_STB_CABOS = 'rebootManualStbCabos-',
    TV_VERIFICAR_CABOS_ETHERNET = 'verificarCabosEthernet-',
    TROCA_NOME = 'trocaNome-',
    TROCA_SENHA = 'trocaSenha-',
    AUTH_HDM = 'authHdm-',
    CORRECAO_DNS = 'correcaoDns-',
    MODEM_CONFIG = 'configModem-',
    ISOLAR_REDE = 'isolarRede-',
    TS_WIFI = 'tsWifi-',
    TS_CABOS = 'tsCabos-',
    WIFI_SIGNAL = 'wifiSignal-',
    TESTE_VELOCIDADE = 'speedTest-',
    FIXO_RUIDO = 'fixoRuido-',
    FIXO_MUDO = 'fixoMudo-',
    FIXO_NAO_COMPLETA = 'fixoNaoCompleta-',
    FILTRO_E_PROTETOR = 'filtroEProtetor-',
    CABO_RJ11 = 'caboRJ11-',
    CABO_COAXIAL = 'caboCoaxial-',
    SINTONIZAR_CANAL = 'syncChannel-',
    ENVIO_PULSO = 'envioPulso-',
    DIVISOR_SINAL = 'divisorSinal-',
    SINAL_SATELITE = 'sinalSatelite-',
    TV_TELA_PRETA = 'telaPreta-',
    TV_CARTAO_INCOMPATIVEL = 'tvIncompativel-',
    ABERTURA_BD = 'aberturaBD-',
    CONFIG_MODEM_MANUAL = 'configModemManual-',
    ISOLAR_STB = 'isolarStb-',
    COMPATIBILIDADE = 'compatibilidade-',
    COMPATIBILIDADE_MANUAL = 'compatibilidadeManual-',
    LINHA_CRUZADA = 'linhaCruzada-',
    TS_MICROFILTRO = 'microfiltro-',
    VOICENET = 'voicenet-',
    RETIRADA_EQUIPAMENTO_FIBRA = 'RetiradaEquipamentoFibra-',
    DEFEITO_NO_CONTROLE = 'DefeitoNoControle-',
    FIBRA_TV_IMAGEM_TOMADA = 'fibraTvImagemTomada-',
    FIBRA_NETFLIX = 'fibraNetflix-',
    ABERTURA_BD_CORP = 'aberturaBDCorp-',
    FIND_STB = 'findStb-'
}

export enum BlockTypes {
    FINANCEIRO = 'FINANCEIRO',
    FALHA_MASSIVA = 'FALHA_MASSIVA',
    REPARO = 'REPARO',
    OS = 'OS',
    LINHA = 'LINHA',
}